+++ frontend/.next/server/edge-runtime-webpack.js
@@ -131,7 +131,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-/******/ 		__webpack_require__.h = () => ("bfddfcf9f5d99265")
+/******/ 		__webpack_require__.h = () => ("d21855aaf78d84f3")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/global */