+++ frontend/.next/static/chunks/webpack.js
@@ -192,7 +192,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	!function() {
-/******/ 		__webpack_require__.h = function() { return "c5e6e2e9e426030c"; }
+/******/ 		__webpack_require__.h = function() { return "5d309f002ae5e26e"; }
 /******/ 	}();
 /******/ 	
 /******/ 	/* webpack/runtime/global */